pyproject.toml
[tool.poetry]
name = "multi-pass"
version = "1.0a3"
description = "multi-pass is a tool to transparently work with multiple pass repositories."
authors = ["Julian Fölsch <julianf@paktosan.de>"]
packages = [
    { include = "multipass" },
    ]
readme = "README.md"
license = "GPL-3.0-or-later"
repository = "https://github.com/Paktosan/multipass"
classifiers = ["Environment :: Console", "Development Status :: 2 - Pre-Alpha", "Intended Audience :: End Users/Desktop",
"Operating System :: Unix", "Topic :: Security", "Topic :: Utilities"]

[tool.poetry.dependencies]
python = "^3.10"
docopt-ng = "^0.8.1"

[tool.poetry.scripts]
multi-pass = 'multipass.cli:run'


[tool.poe.tasks]
test = "pytest"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

